Require backports.lzma on Python 2.7
The `backports.lzma` package is an optional dependency on Python 2.7 for
reading compressed TIFF files. It's mainly optional as it's a little
tricky getting `backports.lzma` built (particularly on Windows). However
as we do have a copy of `backports.lzma` on all platforms in
conda-forge, it makes sense to make it a requirement on Python 2.7 to
the ensure full functionality of `tifffile`.